backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.2
google-generativeai>=0.8.0
pydantic>=2.10.0
python-multipart>=0.0.9
playwright>=1.48.0
python-dotenv>=1.0.0
